wahBot/requirements.txt
python-telegram-bot~=12.1.1
